Parse file for any number of CPUs.
